feat: opt-in consolidate outputs
